# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Retail POS System",
        "title": "Retail POS System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9095",
    "basePath": "/api",
    "paths": {
...
        "/pricing-rules": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of pricing rules with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pricing"
                ],
                "summary": "List pricing rules",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search rule name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by product ID",
                        "name": "product_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by type (special_price|promotion)",
                        "name": "pricing_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by method (fixed_price|discount_percent|discount_amount|markup_percent)",
                        "name": "pricing_method",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by brand ID",
                        "name": "brand_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by customer group ID",
                        "name": "customer_group_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by store ID",
                        "name": "store_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by active status",
                        "name": "is_active",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by approval status (draft|pending|approved|rejected)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new pricing rule. Requires at least one target (product_id, category_id, or brand_id).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pricing"
                ],
                "summary": "Create a pricing rule",
                "parameters": [
                    {
                        "description": "Pricing rule data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pricing.Rule"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/pricing-rules/check-conflicts": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Check if a proposed rule would conflict with existing active rules",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pricing"
                ],
                "summary": "Check for conflicting pricing rules",
                "parameters": [
                    {
                        "description": "Rule to check",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pricing.checkConflictsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/pricing-rules/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a single pricing rule by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pricing"
                ],
                "summary": "Get pricing rule by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing pricing rule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pricing"
                ],
                "summary": "Update a pricing rule",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pricing.Rule"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a pricing rule by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pricing"
                ],
                "summary": "Delete a pricing rule",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/pricing-rules/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Transition a pending rule to approved status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pricing"
                ],
                "summary": "Approve a pricing rule",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/pricing-rules/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Transition a pending rule to rejected status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pricing"
                ],
                "summary": "Reject a pricing rule",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/pricing-rules/{id}/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Transition a draft rule to pending status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pricing"
                ],
                "summary": "Submit rule for approval",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pricing.Method": {
            "type": "string",
            "enum": [
                "fixed_price",
                "discount_percent",
                "discount_amount",
                "markup_percent"
            ],
            "x-enum-varnames": [
                "PricingMethodFixedPrice",
                "PricingMethodDiscountPct",
                "PricingMethodDiscountAmt",
                "PricingMethodMarkupPct"
            ]
        },
...
        "pricing.Type": {
            "type": "string",
            "enum": [
                "normal",
                "special_price",
                "promotion"
            ],
            "x-enum-comments": {
                "PricingTypeNormal": "fallback only — not creatable"
            },
            "x-enum-descriptions": [
                "fallback only — not creatable",
                "",
                ""
            ],
            "x-enum-varnames": [
                "PricingTypeNormal",
                "PricingTypeSpecialPrice",
                "PricingTypePromotion"
            ]
        },
        "pricing.checkConflictsRequest": {
            "type": "object",
            "required": [
                "pricing_method",
                "pricing_type"
            ],
            "properties": {
                "brand_id": {
                    "type": "integer"
                },
                "category_id": {
                    "type": "integer"
                },
                "exclude_id": {
                    "type": "integer"
                },
                "maximum_quantity": {
                    "type": "integer"
                },
                "minimum_quantity": {
                    "type": "integer"
                },
                "pricing_method": {
                    "$ref": "#/definitions/pricing.Method"
                },
                "pricing_type": {
                    "$ref": "#/definitions/pricing.Type"
                },
                "pricing_value": {
                    "type": "number"
                },
                "priority": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}